RM-926: File plan filters (Hold, Transfer, Unfiled) may not show up to date results when Rm configured for SOLR
  * transfer and hold containers created under file plan
  * holds and transfers now contained within these containers
  * holds named using folder level counter to avoid name clash in container
  * hold and transfer filters now API driven .. so canned query driven .. so *live*
  * unit test updates
  * UI tested



git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/modules/recordsmanagement/HEAD@55091 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ rm-server/source/java/org/alfresco/module/org_alfresco_module_rm/security/FilePlanPermissionServiceImpl.java
@@ -454,19 +454,25 @@ public class FilePlanPermissionServiceImpl implements FilePlanPermissionService,
      */
     private void setPermissionDown(NodeRef nodeRef, String authority, String permission)
     {
-        setPermissionImpl(nodeRef, authority, permission);
-        if (filePlanService.isFilePlanContainer(nodeRef) == true ||
-            recordsManagementService.isRecordFolder(nodeRef) == true)
+        // skip out node's that inherit (for example hold and transfer)
+        if (permissionService.getInheritParentPermissions(nodeRef) == false)
         {
-            List<ChildAssociationRef> assocs = nodeService.getChildAssocs(nodeRef, ContentModel.ASSOC_CONTAINS, RegexQNamePattern.MATCH_ALL);
-            for (ChildAssociationRef assoc : assocs)
-            {
-                NodeRef child = assoc.getChildRef();
-                if (filePlanService.isFilePlanContainer(child) == true ||
-                    recordsManagementService.isRecordFolder(child) == true ||
-                    recordService.isRecord(child) == true)
+            // set permissions
+            setPermissionImpl(nodeRef, authority, permission);
+            
+            if (filePlanService.isFilePlanContainer(nodeRef) == true ||
+                recordsManagementService.isRecordFolder(nodeRef) == true)
+            {                            
+                List<ChildAssociationRef> assocs = nodeService.getChildAssocs(nodeRef, ContentModel.ASSOC_CONTAINS, RegexQNamePattern.MATCH_ALL);
+                for (ChildAssociationRef assoc : assocs)
                 {
-                    setPermissionDown(child, authority, permission);
+                    NodeRef child = assoc.getChildRef();
+                    if (filePlanService.isFilePlanContainer(child) == true ||
+                        recordsManagementService.isRecordFolder(child) == true ||
+                        recordService.isRecord(child) == true)
+                    {
+                        setPermissionDown(child, authority, permission);
+                    }
                 }
             }
         }
@@ -499,21 +505,25 @@ public class FilePlanPermissionServiceImpl implements FilePlanPermissionService,
         {
             public Boolean doWork() throws Exception
             {
-                // Delete permission on this node
-                permissionService.deletePermission(nodeRef, authority, permission);
-
-                if (filePlanService.isFilePlanContainer(nodeRef) == true ||
-                    recordsManagementService.isRecordFolder(nodeRef) == true)
+                // can't delete permissions if inherited (eg hold and transfer containers)
+                if (permissionService.getInheritParentPermissions(nodeRef) == false)
                 {
-                    List<ChildAssociationRef> assocs = nodeService.getChildAssocs(nodeRef, ContentModel.ASSOC_CONTAINS, RegexQNamePattern.MATCH_ALL);
-                    for (ChildAssociationRef assoc : assocs)
+                    // Delete permission on this node
+                    permissionService.deletePermission(nodeRef, authority, permission);
+    
+                    if (filePlanService.isFilePlanContainer(nodeRef) == true ||
+                        recordsManagementService.isRecordFolder(nodeRef) == true)
                     {
-                        NodeRef child = assoc.getChildRef();
-                        if (filePlanService.isFilePlanContainer(child) == true ||
-                            recordsManagementService.isRecordFolder(child) == true ||
-                            recordService.isRecord(child) == true)
+                        List<ChildAssociationRef> assocs = nodeService.getChildAssocs(nodeRef, ContentModel.ASSOC_CONTAINS, RegexQNamePattern.MATCH_ALL);
+                        for (ChildAssociationRef assoc : assocs)
                         {
-                            deletePermission(child, authority, permission);
+                            NodeRef child = assoc.getChildRef();
+                            if (filePlanService.isFilePlanContainer(child) == true ||
+                                recordsManagementService.isRecordFolder(child) == true ||
+                                recordService.isRecord(child) == true)
+                            {
+                                deletePermission(child, authority, permission);
+                            }
                         }
                     }
                 }
